+++ assets/javascripts/kitten/index.js
@@ -217,6 +217,7 @@ export { UsersIconNext } from './components/graphics/icons-next/users-icon-next'
 export { WhatsappIconNext } from './components/graphics/icons-next/whatsapp-icon-next'
 export { WindowEditIconNext } from './components/graphics/icons-next/window-edit-icon-next'
 export { HeartIconNext } from './components/graphics/icons-next/heart-icon-next'
+export { LargeArrowIconNext } from './components/graphics/icons-next/large-arrow-icon-next'
 
 // Graphics/Illustrations
 export { LightbulbIllustration } from './components/graphics/illustrations/lightbulb-illustration'